Updated CSS and refactored code
