+++ client/src/App.css
@@ -25,18 +25,24 @@ nav img {
   list-style: none;
 }
 
-/* Collect Form Styling */
+@media only screen and (max-width: 749px) {
+  .nav-links {
+    justify-content: space-between;
+  }
+}
+
+/* Form Styling */
 * {
   font-family: "Helvetica Neue", Helvetica;
   font-size: 15px;
   font-variant: normal;
-  padding: 0;
-  margin: 0;
+  /* padding: 0;
+  margin: 0; */
 }
 
-html {
+/* html {
   height: 100%;
-}
+} */
 
 body {
   background: rgba(70, 117, 153);
@@ -45,10 +51,16 @@ body {
 }
 
 form {
-  max-width: 500px;
   min-width: 490px;
   margin: 3%;
   display: inline-block;
+}
+
+@media only screen and (max-width: 749px) {
+  form {
+    width: 90%;
+    min-width: unset;
+  }
 }
 
 iframe {
@@ -119,6 +131,7 @@ button {
   height: 40px;
   line-height: 38px;
   outline: none;
+  cursor:pointer;
 }
 
 button:focus {
