+++ client/src/iFrame.js
@@ -1,0 +1,16 @@
+const css = {
+    fontFamily: '"Helvetica Neue", Helvetica',
+    boxSizing: "border-box",
+    lineHeight: "1.5em",
+    fontSize: "14px",
+    fontWeight: "200",
+    border: "none",
+    color: "#31325F",
+    width: "100%",
+    height: "100%",
+    "&::placeholder": {
+      color: "#CFD7E0"
+    }
+  };
+
+  export default css
